# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: huMANity API - Explore history and culture of nations
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/uploads/finalize:
    post:
      tags: [Storage]
      operationId: finalizeUpload
      summary: Finalize an uploaded object and publish it
      description: |
        Called after a successful PUT to the presigned URL. Sets a public ACL
        policy on the uploaded object (owned by the requesting user) so it can be
        served from the public profile. Requires authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalizeUploadRequest"
      responses:
        "200":
          description: Object finalized and published.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalizedUploadResult"
        "400":
          description: Missing or invalid object path.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to finalize object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /countries/{code}/timeline:
    get:
      operationId: listTimeline
      tags: [timeline]
      summary: Get timeline events for a country
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
    post:
      operationId: createTimelineEvent
      tags: [timeline]
      summary: Add a timeline event
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimelineEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEvent"
...
  /countries/{code}/stories:
    get:
      operationId: listStories
      tags: [stories]
      summary: Get personal stories for a country
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Personal stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Story"
    post:
      operationId: createStory
      tags: [stories]
      summary: Add a personal story
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
...
  /me/profile:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get the signed-in user's profile
      responses:
        "200":
          description: The user's profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
        "404":
          description: Profile not yet created
    put:
      operationId: updateMyProfile
      tags: [profile]
      summary: Create or update the signed-in user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: The saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
  /me/pledge:
    post:
      operationId: pledgeAsUser
      tags: [profile]
      summary: Mark the signed-in user as having taken the Humanity Pledge
      responses:
        "200":
          description: The updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
  /profile/{userId}:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get a public profile by user id
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
...
components:
  schemas:
...
    TimelineEvent:
      type: object
      required: [id, countryCode, year, title, description]
      properties:
        id:
          type: integer
        countryCode:
          type: string
        year:
          type: integer
        title:
          type: string
        description:
          type: string
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
    TimelineEventInput:
      type: object
      required: [year, title, description]
      properties:
        year:
          type: integer
        title:
          type: string
        description:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
        category:
          type: string
...
    Story:
      type: object
      required: [id, countryCode, title, content]
      properties:
        id:
          type: integer
        countryCode:
          type: string
        title:
          type: string
        author:
          type: ["string", "null"]
        content:
          type: string
        audioUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    StoryInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        author:
          type: string
        content:
          type: string
        audioUrl:
          type: string
        videoUrl:
          type: string
        imageUrl:
          type: string
...
    Profile:
      type: object
      required:
        - id
        - userId
        - displayName
        - languages
        - interests
        - favoriteBooks
        - favoriteMusic
        - countriesExplored
        - humanityScore
        - pledged
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
          nullable: true
        displayName:
          type: string
        countryCode:
          type: string
          nullable: true
        countryName:
          type: string
          nullable: true
        countryFlagUrl:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        culturalBackground:
          type: string
          nullable: true
        languages:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        favoriteBooks:
          type: array
          items:
            type: string
        favoriteMusic:
          type: array
          items:
            type: string
        profileSong:
          type: string
          nullable: true
        profileSongTitle:
          type: string
          nullable: true
        profileSongArtist:
          type: string
          nullable: true
        profileSongArtwork:
          type: string
          nullable: true
        profileSongPreviewUrl:
          type: string
          nullable: true
        profileSongUrl:
          type: string
          nullable: true
        countriesExplored:
          type: array
          items:
            type: string
        humanityScore:
          type: integer
        pledged:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    ProfileInput:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 1
        username:
          type: string
          nullable: true
        countryCode:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        culturalBackground:
          type: string
          nullable: true
        languages:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        favoriteBooks:
          type: array
          items:
            type: string
        favoriteMusic:
          type: array
          items:
            type: string
        profileSong:
          type: string
          nullable: true
        profileSongTitle:
          type: string
          nullable: true
        profileSongArtist:
          type: string
          nullable: true
        profileSongArtwork:
          type: string
          nullable: true
        profileSongPreviewUrl:
          type: string
          nullable: true
        profileSongUrl:
          type: string
          nullable: true
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    FinalizeUploadRequest:
      type: object
      required: [objectPath]
      properties:
        objectPath:
          type: string
          minLength: 1
          description: Normalized object path returned by request-url (e.g. `/objects/uploads/uuid`).
    FinalizedUploadResult:
      type: object
      required: [objectPath]
      properties:
        objectPath:
          type: string
          description: Normalized object path, now publicly served.
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...